# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Rent a Tool API",
    "description": "API documentation for the Rent a Tool platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
      "ProfileSummaryDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "object"
          },
          "firstName": {
            "type": "object"
          },
          "lastName": {
            "type": "object"
          },
          "phoneCode": {
            "type": "object"
          },
          "phoneNumber": {
            "type": "object"
          }
        }
      },
      "AddressSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "label": {
            "type": "object"
          },
          "street": {
            "type": "object"
          },
          "city": {
            "type": "object"
          },
          "country": {
            "type": "object"
          }
        },
        "required": [
          "id"
        ]
      },
      "AuthMeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "verificationTier": {
            "type": "string",
            "enum": [
              "UNVERIFIED",
              "TIER_1",
              "TIER_2",
              "TIER_3"
            ]
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileSummaryDto"
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddressSummaryDto"
            }
          }
        },
        "required": [
          "id",
          "email",
          "verificationTier"
        ]
      },
...
    }
  }
}